Make homedir_test windows compatible
